+++ packages/vue/src/Form/ColorPicker/components/ColorAction.vue
@@ -1,0 +1,56 @@
+<template>
+  <div
+    class="ck-cp-controller-bar"
+    style="display: inline-flex; justify-content: end; position: relative"
+  >
+    <button
+      v-if="isEyeDropperUsing && showEyeDrop"
+      id="cp-btn-eyedropper"
+      type="button"
+      class="cp-btn"
+      style="padding: 7px"
+      @click="$emit('onClickEyeDropper')"
+    >
+      <svg
+        version="1.1"
+        xmlns="http://www.w3.org/2000/svg"
+        width="32"
+        height="32"
+        viewBox="0 0 32 32"
+      >
+        <path
+          d="M21.35 1.825l-6.344 6.35-0.588-0.588c-0.781-0.781-2.050-0.781-2.831 0s-0.781 2.050 0 2.831l10 10c0.781 0.781 2.050 0.781 2.831 0s0.781-2.050 0-2.831l-0.587-0.587 6.344-6.35c2.438-2.438 2.438-6.388 0-8.819s-6.387-2.438-8.819 0zM3.462 20.206c-0.938 0.938-1.462 2.212-1.462 3.538v2.65l-1.663 2.494c-0.531 0.794-0.425 1.85 0.25 2.525s1.731 0.781 2.525 0.25l2.494-1.663h2.65c1.325 0 2.6-0.525 3.537-1.462l7.544-7.544-2.831-2.831-7.544 7.544c-0.188 0.188-0.444 0.294-0.706 0.294h-2.256v-2.256c0-0.262 0.106-0.519 0.294-0.706l7.544-7.544-2.831-2.831-7.544 7.544z"
+        ></path>
+      </svg>
+    </button>
+
+    <button
+      v-if="showColorList"
+      type="button"
+      class="cp-btn"
+      style="padding: 7px"
+      @click="$emit('onSaveColor')"
+    >
+      <svg
+        version="1.1"
+        xmlns="http://www.w3.org/2000/svg"
+        width="28"
+        height="32"
+        viewBox="0 0 28 32"
+      >
+        <path
+          d="M4 2c-2.206 0-4 1.794-4 4v20c0 2.206 1.794 4 4 4h20c2.206 0 4-1.794 4-4v-15.169c0-1.063-0.419-2.081-1.169-2.831l-4.831-4.831c-0.75-0.75-1.769-1.169-2.831-1.169h-15.169zM4 8c0-1.106 0.894-2 2-2h12c1.106 0 2 0.894 2 2v4c0 1.106-0.894 2-2 2h-12c-1.106 0-2-0.894-2-2v-4zM14 18c2.209 0 4 1.791 4 4s-1.791 4-4 4v0c-2.209 0-4-1.791-4-4s1.791-4 4-4v0z"
+        ></path>
+      </svg>
+    </button>
+  </div>
+</template>
+
+<script setup>
+defineProps({
+  isEyeDropperUsing: { default: false, type: Boolean },
+  showColorList: { default: true, type: Boolean },
+  showEyeDrop: { default: true, type: Boolean },
+});
+defineEmits(["onSaveColor", "onClickEyeDropper"]);
+</script>
